pyproject.toml
[project]
name = "mafia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.13.3",
    "sqlalchemy==2.0.35",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "asyncpg==0.30.0",
    "fastapi[standard]==0.115.11",
    "uvicorn[standard]>=0.27.0",
    "pyjwt>=2.10.1",
    "bcrypt>=4.3.0",
    "python-multipart>=0.0.20",
]
